+++ lhc_web/modules/lhuser/editgroup.php
@@ -5,30 +5,35 @@ $tpl = erLhcoreClassTemplate::getInstance('lhuser/editgroup.tpl.php');
 $Group = erLhcoreClassUser::getSession()->load( 'erLhcoreClassModelGroup', (int)$Params['user_parameters']['group_id'] );
 
 if (isset($_POST['Update_group']) )
-{    
+{
    $definition = array(
         'Name' => new ezcInputFormDefinitionElement(
             ezcInputFormDefinitionElement::REQUIRED, 'unsafe_raw'
-        )       
+        )
     );
-    
+
     $form = new ezcInputForm( INPUT_POST, $definition );
     $Errors = array();
-    
+
+    if (!isset($_POST['csfr_token']) || !$currentUser->validateCSFRToken($_POST['csfr_token'])) {
+    	erLhcoreClassModule::redirect('user/userlist');
+    	exit;
+    }
+
     if ( !$form->hasValidData( 'Name' ) || $form->Name == '' )
     {
         $Errors[] =  erTranslationClassLhTranslation::getInstance()->getTranslation('user/editgroup','Please enter a group name');
     }
-    
+
     if (count($Errors) == 0)
-    {     
+    {
         $Group->name = $form->Name;
-    
+
         erLhcoreClassUser::getSession()->update($Group);
-                     
+
         erLhcoreClassModule::redirect('user/grouplist');
         exit;
-        
+
     }  else {
         $tpl->set('errors',$Errors);
     }
@@ -36,19 +41,29 @@ if (isset($_POST['Update_group']) )
 
 if (isset($_POST['AssignRoles']) && isset($_POST['RoleID']) && count($_POST['RoleID']) > 0)
 {
+	if (!isset($_POST['csfr_token']) || !$currentUser->validateCSFRToken($_POST['csfr_token'])) {
+		erLhcoreClassModule::redirect('user/userlist');
+		exit;
+	}
+
     foreach ($_POST['RoleID'] as $RoleID)
-    {                
-        $GroupRole = new erLhcoreClassModelGroupRole();        
+    {
+        $GroupRole = new erLhcoreClassModelGroupRole();
         $GroupRole->group_id = $Group->id;
-        $GroupRole->role_id = $RoleID;        
+        $GroupRole->role_id = $RoleID;
         erLhcoreClassRole::getSession()->save($GroupRole);
     }
 }
 
 if (isset($_POST['Remove_user_from_group']) && isset($_POST['AssignedID']) && count($_POST['AssignedID']) > 0)
 {
+	if (!isset($_POST['csfr_token']) || !$currentUser->validateCSFRToken($_POST['csfr_token'])) {
+		erLhcoreClassModule::redirect('user/userlist');
+		exit;
+	}
+
     foreach ($_POST['AssignedID'] as $AssignedID)
-    {                
+    {
         $group_user = erLhcoreClassModelGroupUser::fetch($AssignedID);
         $group_user->removeThis();
     }
@@ -56,8 +71,13 @@ if (isset($_POST['Remove_user_from_group']) && isset($_POST['AssignedID']) && co
 
 if (isset($_POST['Remove_role_from_group']) && isset($_POST['AssignedID']) && count($_POST['AssignedID']) > 0)
 {
+	if (!isset($_POST['csfr_token']) || !$currentUser->validateCSFRToken($_POST['csfr_token'])) {
+		erLhcoreClassModule::redirect('user/userlist');
+		exit;
+	}
+
     foreach ($_POST['AssignedID'] as $AssignedID)
-    {                
+    {
         erLhcoreClassGroupRole::deleteGroupRole($AssignedID);
     }
 }
